requirements.txt
Werkzeug
python-dotenv
mysql-connector-python
spyne
